+++ src/node_dependencies/resolve.js
