+++ src/zip_node.js
